Split out WriteBuffer and derived classes into a separate writebuf.cc/h file
